correct list logic

+++ src/list/config_list.h
@@ -7,14 +7,16 @@
 typedef Word *elem_t;
 
 
-const elem_t Poison_val = nullptr;  //<- In free nodes is written
+const elem_t Poison_val  = nullptr;  //<- In free nodes is written
 
-const int Poison_ptr  = -126;   //<- Written to a pointer when the list is cleared
+const size_t Poison_ptr  = -126;   //<- Written to a pointer when the list is cleared
 
-const int Invalid_ind = 0;
+const size_t Invalid_ind = 0;
+
 
 #define USE_LOG                 //<- connect when we use logs
 
+//#define CHECK_LIST 
 
 #define LIST_DATA_CHECK         //<- Checking non-free list nodes for correct transitions and values
 
@@ -23,5 +25,9 @@ const int Invalid_ind = 0;
 //#define GRAPH_DUMP
 
 #define ELEM_T_SPEC "p"         //<- specifier character to print elem
+
+
+//#define OPTIMIZE_FIND_VER
+
 
 #endif  //endif _LIST_CONFIG_H_
